Switched PointsScored to decimal

+++ Source/BusinessLogic/Models/Games/PlayerRank.cs
@@ -26,7 +26,7 @@ namespace BusinessLogic.Models.Games
         public int PlayerId { get; set; }
         [Required]
         public int GameRank { get; set; }
-        public int? PointsScored { get; set; }
+        public decimal? PointsScored { get; set; }
     }
 
     public interface IPlayerRank
